Rework TouchRunner into something that doesn't need UIKit nor MonoTouch.Dialog.
So that it can run on WatchOS.

This is done by creating a base class that doesn't need UIKit / MonoTouch.Dialog,
and for iOS/tvOS, the derived class has all the UI code, and for watchOS
(in a different derive class) there is no UI code.

Also add the possibility if passing arguments through environment variables,
since we can't pass arguments when launching watchOS apps, only environment
variables.