[TouchRunner] Don't auto-exit if we couldn't connect to the test runner.
Extensions will automatically re-launch if they exit manually (or if they
crash), which means that every time a test run from an extension terminates,
iOS will re-launch it.

Unfortunately iOS will launch the extension with the same environment
variables, thus configuring the process to autorun / autoexit, which leads to
an infinite number of test runs (except that iOS detects extensions that
frequently exits and invalidates them after a while).

Fortunately, this second time we won't be able to connect to the test runner, so
we can use that to break the cycle.